Add files via upload
In this project I developed, I made an advanced To-Do List. To-Do added in this To-Do List is added under the "Current To Do" heading. If desired, an update can be made by editing this To Do. If it is deleted, this To-Do is sent under the "Deleted" heading, which has the logic of "Recycle Bin". This To-Do can be returned to the "Current To-Do" area again if desired. In case of deletion, it is completely deleted permanently. In addition, certain conditions are filtered in To Do's titled "Current To-Do" and "Deleted". As a result of this filtering, if there is no To Do with the sought value, a warning message appears. If any, filtered To Dos can be deleted thanks to the "Delete Filtered To Dos" button and sent to the "Deleted" heading from the "Current To Dos" heading. In addition to these, by clicking on the To Dos under the "Current To Do" heading, which are completed but do not want to be deleted, the background and text color are changed to provide a distinction. It turns out that To-Do is complete before it needs to be deleted. In addition to all these, the To-Dos under the selected title are instantly deleted thanks to the "Clean all tasks" and "Clean all deleted tasks" buttons. As with normal deletion, To-Dos under the "Current To Do" heading are sent to the "Deleted" heading. Thank you, for your interest.